- Refactored the generator so that first it updates the game data, then it creates the patches. To help reduce the complexity of the generator function and to help make sure we don't generate patches that rely on game data that still needs to be updated.
