--- src/main/resources/patchSpecs/contestEncounterSlot.yml
@@ -1,3 +1,0 @@
-description: "Data for a bug catching contest encounter slot."
-values:
-  data: "$encounterRate $pokemonId $minLevel $maxLevel"
