--- src/main/resources/patchSpecs/fishingTimeGroup.yml
@@ -1,3 +1,0 @@
-description: "Data for a fishing time group."
-values:
-  data: "$dayPokemonId $dayLevel $nightPokemonId $nightLevel"
