--- src/main/resources/patchSpecs/levelUpMove.yml
@@ -1,3 +1,0 @@
-description: "Defines a level up move."
-values: 
-  data: "$level $moveId"
