+++ src/main/resources/patchSpecs/pokemonEvolutionsAndLevelUpMoves.yml
@@ -1,3 +1,3 @@
 description: "Defines a pokemon's evolutions and level up moves."
 values: 
-  data: "#evolutions.$data 00 #levelUpMoves.$data 00"
+  data: "$data"
