--- src/main/resources/patchSpecs/wildEncounters.yml
@@ -1,22 +1,0 @@
-description: "Changes the data for the pokemon that can be randomly encountered in the wild."
-changes:
-- description: "Changes the data for the pokemon that can be randomly encountered in the tall grass, or surfing on the water."
-  locations:
-  - bank: 10
-    address: 0x65E9
-  value: "#johtoLandEncounterGroups.$data FF #johtoWaterEncounterGroups.$data FF #kantoLandEncounterGroups.$data FF #kantoWaterEncounterGroups.$data FF #swarmEncounterGroups.$data FF"
-- description: "Changes the data for the pokemon that can be encountered in the bug catching contest."
-  locations:
-  - bank: 37
-    address: 0x7D87
-  value: "#contestEncounterSlots.$data"
-- description: "Changes the data for the pokemon that can be encountered fishing."
-  locations:
-  - bank: 36
-    address: 0x64E3
-  value: "#fishingEncounterSlots.$data #fishingTimeEncounterGroups.$data"
-- description: "Changes the data for the pokemon that can be encountered in trees and rocks."
-  locations:
-  - bank: 46
-    address: 0x42FA
-  value: "#treeEncounterGroups.$data #rockEncounterSlots.$data FF"
