+++ src/shared/types/gameData/eventPokemon.ts
@@ -1,0 +1,34 @@
+import type { EventPokemonId } from "@shared/types/gameDataIds/eventPokemon"
+import type { PokemonId } from "@shared/types/gameDataIds/pokemon"
+
+export const eventPokemonMap: Record<EventPokemonId, PokemonId> = {
+  RATTATA: "RATTATA",
+  SUDOWOODO: "SUDOWOODO",
+  RAIKOU: "RAIKOU",
+  ENTEI: "ENTEI",
+  SUICUNE: "SUICUNE",
+  GYARADOS: "GYARADOS",
+  VOLTORB: "VOLTORB",
+  GEODUDE: "GEODUDE",
+  KOFFING: "KOFFING",
+  ELECTRODE1: "ELECTRODE",
+  ELECTRODE2: "ELECTRODE",
+  ELECTRODE3: "ELECTRODE",
+  LAPRAS: "LAPRAS",
+  SNORLAX: "SNORLAX",
+  HO_OH: "HO_OH",
+  LUGIA: "LUGIA",
+  CELEBI: "CELEBI",
+  TOGEPI: "TOGEPI",
+  SPEAROW: "SPEAROW",
+  SHUCKLE: "SHUCKLE",
+  EEVEE: "EEVEE",
+  DRATINI: "DRATINI",
+  TYROGUE: "TYROGUE",
+  ABRA: "ABRA",
+  CUBONE: "CUBONE",
+  WOBBUFFET: "WOBBUFFET",
+  PIKACHU: "PIKACHU",
+  PORYGON: "PORYGON",
+  LARVITAR: "LARVITAR",
+}
