+++ src/shared/types/gameDataIds/eventPokemon.ts
@@ -1,0 +1,33 @@
+export const eventPokemonIds = [
+  "RATTATA",
+  "SUDOWOODO",
+  "RAIKOU",
+  "ENTEI",
+  "SUICUNE",
+  "GYARADOS",
+  "VOLTORB",
+  "GEODUDE",
+  "KOFFING",
+  "ELECTRODE1",
+  "ELECTRODE2",
+  "ELECTRODE3",
+  "LAPRAS",
+  "SNORLAX",
+  "HO_OH",
+  "LUGIA",
+  "CELEBI",
+  "TOGEPI",
+  "SPEAROW",
+  "SHUCKLE",
+  "EEVEE",
+  "DRATINI",
+  "TYROGUE",
+  "ABRA",
+  "CUBONE",
+  "WOBBUFFET",
+  "PIKACHU",
+  "PORYGON",
+  "LARVITAR",
+] as const
+
+export type EventPokemonId = typeof eventPokemonIds[number]
